[MAPSME-5374] Updated intermediateIndex type.

+++ iphone/Maps/Core/Routing/MWMRoutePoint+CPP.h
@@ -9,12 +9,12 @@
 
 - (instancetype)initWithURLSchemeRoutePoint:(url_scheme::RoutePoint const &)point
                                        type:(MWMRoutePointType)type
-                          intermediateIndex:(int8_t)intermediateIndex;
+                          intermediateIndex:(size_t)intermediateIndex;
 - (instancetype)initWithRouteMarkData:(RouteMarkData const &)point;
 - (instancetype)initWithPoint:(m2::PointD const &)point
                         title:(NSString *)title
                      subtitle:(NSString *)subtitle
                          type:(MWMRoutePointType)type
-            intermediateIndex:(int8_t)intermediateIndex;
+            intermediateIndex:(size_t)intermediateIndex;
 
 @end
